+++ src/roq/samples/example-3/strategy.cpp
@@ -2,6 +2,8 @@
 
 #include "roq/samples/example-3/strategy.h"
 
+#include <absl/flags/flag.h>
+
 #include <limits>
 
 #include "roq/logging.h"
@@ -14,8 +16,10 @@ namespace samples {
 namespace example_3 {
 
 Strategy::Strategy(client::Dispatcher &dispatcher)
-    : dispatcher_(dispatcher),
-      instrument_(FLAGS_exchange, FLAGS_symbol, FLAGS_account) {
+    : dispatcher_(dispatcher), instrument_(
+                                   absl::GetFlag(FLAGS_exchange),
+                                   absl::GetFlag(FLAGS_symbol),
+                                   absl::GetFlag(FLAGS_account)) {
 }
 
 void Strategy::operator()(const Event<Timer> &event) {
@@ -25,7 +29,8 @@ void Strategy::operator()(const Event<Timer> &event) {
   if (next_sample_.count())  // initialized?
     update_model();
   auto now = std::chrono::duration_cast<std::chrono::seconds>(event.value.now);
-  next_sample_ = now + std::chrono::seconds{FLAGS_sample_freq_secs};
+  next_sample_ =
+      now + std::chrono::seconds{absl::GetFlag(FLAGS_sample_freq_secs)};
   // possible extension: reset request timeout
 }
 
@@ -114,7 +119,7 @@ void Strategy::update_model() {
 }
 
 void Strategy::try_trade(Side side, double price) {
-  if (FLAGS_enable_trading == false) {
+  if (absl::GetFlag(FLAGS_enable_trading) == false) {
     LOG(WARNING)("Trading *NOT* enabled");
     return;
   }
@@ -129,7 +134,10 @@ void Strategy::try_trade(Side side, double price) {
     if (side != working_side_) {
       LOG(INFO)("*** CANCEL WORKING ORDER ***");
       dispatcher_.send(
-          CancelOrder{.account = FLAGS_account, .order_id = working_order_id_},
+          CancelOrder{
+              .account = absl::GetFlag(FLAGS_account),
+              .order_id = working_order_id_,
+          },
           uint8_t{0});
     }
     return;
@@ -141,10 +149,10 @@ void Strategy::try_trade(Side side, double price) {
   auto order_id = ++max_order_id_;
   dispatcher_.send(
       CreateOrder{
-          .account = FLAGS_account,
+          .account = absl::GetFlag(FLAGS_account),
           .order_id = order_id,
-          .exchange = FLAGS_exchange,
-          .symbol = FLAGS_symbol,
+          .exchange = absl::GetFlag(FLAGS_exchange),
+          .symbol = absl::GetFlag(FLAGS_symbol),
           .side = side,
           .quantity = instrument_.min_trade_vol(),
           .order_type = OrderType::LIMIT,